fixed warning about printf format specifications <-> argument mismatches in wxLogLastError
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@15960 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/log.h
@@ -577,14 +577,14 @@ void WXDLLEXPORT wxSafeShowMessage(const wxString& title, const wxString& text);
 #ifdef __VISUALC__
     #define wxLogApiError(api, rc)                                            \
         wxLogDebug(wxT("%s(%d): '%s' failed with error 0x%08lx (%s)."),       \
-                   __TFILE__, __LINE__, api,                              \
-                   rc, wxSysErrorMsg(rc))
+                   __TFILE__, __LINE__, api,                                  \
+                   (long)rc, wxSysErrorMsg(rc))
 #else // !VC++
     #define wxLogApiError(api, rc)                                            \
         wxLogDebug(wxT("In file %s at line %d: '%s' failed with "             \
-                      "error 0x%08lx (%s)."),                                 \
-                   __TFILE__, __LINE__, api,                              \
-                   rc, wxSysErrorMsg(rc))
+                       "error 0x%08lx (%s)."),                                \
+                   __TFILE__, __LINE__, api,                                  \
+                   (long)rc, wxSysErrorMsg(rc))
 #endif // VC++/!VC++
 
     #define wxLogLastError(api) wxLogApiError(api, wxSysErrorCode())
